refactor(frontend): decouple and modularize Productivity Screen
- Extract state management and business logic into \`useProductivityViewModel\`
- Decompose complex UI into \`ProductivityHeader\`, \`ProductivityTabs\`, \`TaskPriorityFilters\`, and \`ProductivityEmptyState\`
- Extract \`renderItem\` outside the main component for strict memoization
- Apply SOLID principles to improve readability and maintainability
